Create Uber.js
Here is a framework for adding the Uber capabilities to the App. I have 3 functions so far. The get price and time estimate are pretty straightforward. The functions are sending html requests to the uber api and parsing the json data to be returned as a price in dollars / time in seconds. The callUber is function launches uber from the app if it is installed on the user's device . Otherwise it will redirect the user to the mobile website to sign or whatever we decide to do. The functions are not complete and the only one I know somewhat works is the first half of the callUber function. So, anyone who wants to take a look at the TODO comments or have any suggestions improving what I have so far be my guest.